fixup! squash! Fix #13073: make CRTICIAL errors cause non-zero exit
